fix: decimal zzzz

+++ Ecommerce-microsservicos/Serviços/Carrinho/Carrinho.API/Basket/StoreBasket/StoreBasketCommandHandler.cs
@@ -32,7 +32,7 @@ namespace Basket.API.Basket.StoreBasket
             foreach(var item in cart.Items)
             {
                 var discountResponse = await discountProto.GetDiscountAsync(new GetDiscountRequest { ProductName = item.ProductName }, cancellationToken: cancellationToken);
-                item.Price -= discountResponse.Amount;
+                item.Price -= (decimal)discountResponse.Amount;
             }
         }
     }
